+++ lib/src/clojure_lsp/api.clj
@@ -254,6 +254,9 @@
 
   `:project-root` a java.io.File representing the project root.
 
+  `:analysis` a map with options on how clojure-lsp should analyze the project, available values are:
+    `:type` keyword values: `project-only` or `project-and-dependencies`. Default to `project-only`.
+
   `:output` a map with options on how the result should be printed, available values are:
     `:format` a keyword specifying in which format the data should be returned, defaults to `:edn`.
     `:filter-keys` a list of keywords in case you want only specific fields from output.