+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   - Show better icons for multimethods, var-arg fns, protocols, records, interfaces and types on `workspace/symbol` and `textDocument/documentSymbol`.
   - Inlining the last binding of a let removes the let. #210
 
+- API/CLI
+  - Allow specify analysis type for `dump` command, default to analyze project only analysis. #1383
+
 ## 2022.11.03-00.14.57
 
 - General